Refactored MainActivity in compose example
