+++ buildSrc/src/main/kotlin/Versions.kt
@@ -13,12 +13,12 @@ import org.gradle.plugin.use.PluginDependencySpec
  */
 object Versions {
     const val org_jetbrains_kotlinx_kotlinx_serialization: String = "0.20.0"
-    // available: "0.20.0-1.4-M1-release-99"
+             // available: "0.20.0-1.4-M1-release-99"
 
     const val org_jetbrains_kotlinx_kotlinx_coroutines: String = "1.3.5"
-    // available: "1.3.5-1.4-M1-release-99"
+             // available: "1.3.5-1.4-M1-release-99"
 
-    const val io_github_reactivecircus_flowbinding: String = "0.10.2"
+    const val io_github_reactivecircus_flowbinding: String = "0.11.1"
 
     const val org_jetbrains_kotlinx: String = "0.14.3"
 
@@ -26,9 +26,9 @@ object Versions {
 
     const val androidx_fragment: String = "1.2.4"
 
-    const val androidx_ui: String = "0.1.0-dev10"
-
     const val androidx_test: String = "1.2.0"
+
+    const val androidx_ui: String = "0.1.0-dev10"
 
     const val com_android_tools_build_gradle: String = "4.1.0-alpha08"
 
@@ -42,13 +42,15 @@ object Versions {
 
     const val lifecycle_runtime_ktx: String = "2.2.0"
 
+    const val compose_compiler: String = "0.1.0-dev10"
+
     const val constraintlayout: String = "2.0.0-beta4"
 
     const val flow_extensions: String = "0.0.4"
 
     const val espresso_core: String = "3.2.0"
 
-    const val lint_gradle: String = "26.6.1" // available: "26.6.3"
+    const val lint_gradle: String = "27.1.0-alpha08"
 
     const val appcompat: String = "1.1.0"
 
@@ -60,12 +62,12 @@ object Versions {
 
     const val ktlint: String = "0.36.0"
 
-    const val aapt2: String = "3.6.1-6040484" // available: "3.6.3-6040484"
+    const val aapt2: String = "4.1.0-alpha08-6376216"
 
-    const val mockk: String = "1.9.3"
+    const val mockk: String = "1.10.0"
 
     /**
-     * Current version: "5.6.4"
+     * Current version: "6.3"
      * See issue 19: How to update Gradle itself?
      * https://github.com/jmfayard/buildSrcVersions/issues/19
      */
@@ -78,4 +80,4 @@ object Versions {
  */
 val PluginDependenciesSpec.buildSrcVersions: PluginDependencySpec
     inline get() =
-        id("de.fayard.buildSrcVersions").version(Versions.de_fayard_buildsrcversions_gradle_plugin)
+            id("de.fayard.buildSrcVersions").version(Versions.de_fayard_buildsrcversions_gradle_plugin)
